feat: make terminal renderer selectable
Add terminal.renderer config with xterm.js as the default and ghostty-web as an opt-in renderer. Wire settings API, UI store hydration, split renderer components, and cover settings plus ghostty e2e behavior.
