+++ frontend/openapi/openapi.yaml
@@ -2101,8 +2101,11 @@ components:
       properties:
         font_family:
           type: string
+        renderer:
+          type: string
       required:
         - font_family
+        - renderer
       type: object
     UpdateSettingsRequest:
       additionalProperties: false
