+++ frontend/package.json
@@ -20,6 +20,9 @@
   "dependencies": {
     "@lucide/svelte": "1.3.0",
     "@middleman/ui": "workspace:*",
+    "@xterm/addon-fit": "^0.11.0",
+    "@xterm/addon-webgl": "^0.19.0",
+    "@xterm/xterm": "^6.0.0",
     "dompurify": "3.3.3",
     "ghostty-web": "^0.4.0",
     "js-toml": "^1.0.3",
